chane version_syntax_converter class to be module at rakelib dir
